Refactor: factor out sampler.rs
